+++ packages/components/pages/f-account-info/CHANGELOG.md
@@ -4,8 +4,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-Latest – to be added to the next release
+v0.14.1
 ------------------------------
+*January 13, 2022*
+
 ### Added
 - Visual tests
 - component tests in `f-account-info.component.spec`